create comment api

+++ php/GoBlog/src/resources/views/js/auth.js
@@ -1,26 +1,30 @@
 import CardAccount from "./template/card_account_template.js";
+import { BASE_API_URL } from "./constant.js";
 
-async function getFollow(id) {
-    const {data} = await $.getJSON(`http://localhost:8000/src/api/get-follow.php?id=${id}`)
-    return data
-}
+$(document).ready(function () {
 
-$(document).ready( async function () {
+    getFollow()
+
+})
+
+$("#main-row").on('click', $("#card-account #logout"), function (event) {
+    if ($(event.target).attr('id') === "logout") {
+        sessionStorage.clear()
+        location.reload()
+    }
+})
+
+export async function getFollow() {
     const user = sessionStorage.getItem('auth') ? JSON.parse(sessionStorage.getItem('auth')) : null
 
     if (user != null) {
-        const data = await getFollow(user.id)
+        const { data } = await $.getJSON(`http://${BASE_API_URL}/get-follow.php?id=${user.id}`)
 
+        $("#main-row").children("#card-account").remove()
         $("#main-row").prepend(CardAccount(user, data))
         $(".gedf-main").removeClass("col-md-9")
         $(".gedf-main").addClass("col-md-6")
         $("#auth-nav").hide()
     }
 
-    $("#logout").click(function () {
-        sessionStorage.clear()
-        location.reload()
-    })
-    
-
-}) 
+}
